+++ multisrc/overrides/mangathemesia/atlantisscan/res/mipmap-hdpi/ic_launcher.png
