+++ multisrc/overrides/mangathemesia/atlantisscan/res/mipmap-mdpi/ic_launcher.png
