+++ multisrc/overrides/mangathemesia/atlantisscan/res/mipmap-xhdpi/ic_launcher.png
