+++ multisrc/overrides/mangathemesia/atlantisscan/res/mipmap-xxhdpi/ic_launcher.png
