+++ multisrc/overrides/mangathemesia/atlantisscan/res/mipmap-xxxhdpi/ic_launcher.png
